pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smallstack-runbook"
version = "1.0.0"
description = "A markdown knowledge base app for Django, built for SmallStack"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Everett Michaud" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
]
dependencies = [
    "django-smallstack @ git+https://github.com/emichaud/django-smallstack.git",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/emichaud/smallstack-runbook"
Repository = "https://github.com/emichaud/smallstack-runbook"

[tool.setuptools.packages.find]
include = ["smallstack_runbook*"]

[tool.setuptools.package-data]
smallstack_runbook = [
    "templates/**/*.html",
    "static/**/*",
    "content/*.md",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["E501"]
